// packages/myso/src/grpc/proto/myso/rpc/v2/ledger_service.client.ts
export interface ILedgerServiceClient {
	/**
	 * Query the service for general information about its current state.
	 *
	 * @generated from protobuf rpc: GetServiceInfo(myso.rpc.v2.GetServiceInfoRequest) returns (myso.rpc.v2.GetServiceInfoResponse);
	 */
	getServiceInfo(
		input: GetServiceInfoRequest,
		options?: RpcOptions,
	): UnaryCall<GetServiceInfoRequest, GetServiceInfoResponse>;
	/**
	 * @generated from protobuf rpc: GetObject(myso.rpc.v2.GetObjectRequest) returns (myso.rpc.v2.GetObjectResponse);
	 */
	getObject(
		input: GetObjectRequest,
		options?: RpcOptions,
	): UnaryCall<GetObjectRequest, GetObjectResponse>;
	/**
	 * @generated from protobuf rpc: BatchGetObjects(myso.rpc.v2.BatchGetObjectsRequest) returns (myso.rpc.v2.BatchGetObjectsResponse);
	 */
	batchGetObjects(
		input: BatchGetObjectsRequest,
		options?: RpcOptions,
	): UnaryCall<BatchGetObjectsRequest, BatchGetObjectsResponse>;
	/**
	 * @generated from protobuf rpc: GetTransaction(myso.rpc.v2.GetTransactionRequest) returns (myso.rpc.v2.GetTransactionResponse);
	 */
	getTransaction(
		input: GetTransactionRequest,
		options?: RpcOptions,
	): UnaryCall<GetTransactionRequest, GetTransactionResponse>;
	/**
	 * @generated from protobuf rpc: BatchGetTransactions(myso.rpc.v2.BatchGetTransactionsRequest) returns (myso.rpc.v2.BatchGetTransactionsResponse);
	 */
	batchGetTransactions(
		input: BatchGetTransactionsRequest,
		options?: RpcOptions,
	): UnaryCall<BatchGetTransactionsRequest, BatchGetTransactionsResponse>;
	/**
	 * @generated from protobuf rpc: GetCheckpoint(myso.rpc.v2.GetCheckpointRequest) returns (myso.rpc.v2.GetCheckpointResponse);
	 */
	getCheckpoint(
		input: GetCheckpointRequest,
		options?: RpcOptions,
	): UnaryCall<GetCheckpointRequest, GetCheckpointResponse>;
	/**
	 * @generated from protobuf rpc: GetEpoch(myso.rpc.v2.GetEpochRequest) returns (myso.rpc.v2.GetEpochResponse);
	 */
	getEpoch(
		input: GetEpochRequest,
		options?: RpcOptions,
	): UnaryCall<GetEpochRequest, GetEpochResponse>;
}
/**
 * @generated from protobuf service myso.rpc.v2.LedgerService
 */
export class LedgerServiceClient implements ILedgerServiceClient, ServiceInfo {
	typeName = LedgerService.typeName;
	methods = LedgerService.methods;
	options = LedgerService.options;
	constructor(private readonly _transport: RpcTransport) {}
	/**
	 * Query the service for general information about its current state.
	 *
	 * @generated from protobuf rpc: GetServiceInfo(myso.rpc.v2.GetServiceInfoRequest) returns (myso.rpc.v2.GetServiceInfoResponse);
	 */
	getServiceInfo(
		input: GetServiceInfoRequest,
		options?: RpcOptions,
	): UnaryCall<GetServiceInfoRequest, GetServiceInfoResponse> {
		const method = this.methods[0],
			opt = this._transport.mergeOptions(options);
		return stackIntercept<GetServiceInfoRequest, GetServiceInfoResponse>(
			'unary',
			this._transport,
			method,
			opt,
			input,
		);
	}
	/**
	 * @generated from protobuf rpc: GetObject(myso.rpc.v2.GetObjectRequest) returns (myso.rpc.v2.GetObjectResponse);
	 */
	getObject(
		input: GetObjectRequest,
		options?: RpcOptions,
	): UnaryCall<GetObjectRequest, GetObjectResponse> {
		const method = this.methods[1],
			opt = this._transport.mergeOptions(options);
		return stackIntercept<GetObjectRequest, GetObjectResponse>(
			'unary',
			this._transport,
			method,
			opt,
			input,
		);
	}
	/**
	 * @generated from protobuf rpc: BatchGetObjects(myso.rpc.v2.BatchGetObjectsRequest) returns (myso.rpc.v2.BatchGetObjectsResponse);
	 */
	batchGetObjects(
		input: BatchGetObjectsRequest,
		options?: RpcOptions,
	): UnaryCall<BatchGetObjectsRequest, BatchGetObjectsResponse> {
		const method = this.methods[2],
			opt = this._transport.mergeOptions(options);
		return stackIntercept<BatchGetObjectsRequest, BatchGetObjectsResponse>(
			'unary',
			this._transport,
			method,
			opt,
			input,
		);
	}
	/**
	 * @generated from protobuf rpc: GetTransaction(myso.rpc.v2.GetTransactionRequest) returns (myso.rpc.v2.GetTransactionResponse);
	 */
	getTransaction(
		input: GetTransactionRequest,
		options?: RpcOptions,
	): UnaryCall<GetTransactionRequest, GetTransactionResponse> {
		const method = this.methods[3],
			opt = this._transport.mergeOptions(options);
		return stackIntercept<GetTransactionRequest, GetTransactionResponse>(
			'unary',
			this._transport,
			method,
			opt,
			input,
		);
	}
	/**
	 * @generated from protobuf rpc: BatchGetTransactions(myso.rpc.v2.BatchGetTransactionsRequest) returns (myso.rpc.v2.BatchGetTransactionsResponse);
	 */
	batchGetTransactions(
		input: BatchGetTransactionsRequest,
		options?: RpcOptions,
	): UnaryCall<BatchGetTransactionsRequest, BatchGetTransactionsResponse> {
		const method = this.methods[4],
			opt = this._transport.mergeOptions(options);
		return stackIntercept<BatchGetTransactionsRequest, BatchGetTransactionsResponse>(
			'unary',
			this._transport,
			method,
			opt,
			input,
		);
	}
	/**
	 * @generated from protobuf rpc: GetCheckpoint(myso.rpc.v2.GetCheckpointRequest) returns (myso.rpc.v2.GetCheckpointResponse);
	 */
	getCheckpoint(
		input: GetCheckpointRequest,
		options?: RpcOptions,
	): UnaryCall<GetCheckpointRequest, GetCheckpointResponse> {
		const method = this.methods[5],
			opt = this._transport.mergeOptions(options);
		return stackIntercept<GetCheckpointRequest, GetCheckpointResponse>(
			'unary',
			this._transport,
			method,
			opt,
			input,
		);
	}
	/**
	 * @generated from protobuf rpc: GetEpoch(myso.rpc.v2.GetEpochRequest) returns (myso.rpc.v2.GetEpochResponse);
	 */
	getEpoch(
		input: GetEpochRequest,
		options?: RpcOptions,
	): UnaryCall<GetEpochRequest, GetEpochResponse> {
		const method = this.methods[6],
			opt = this._transport.mergeOptions(options);
		return stackIntercept<GetEpochRequest, GetEpochResponse>(
			'unary',
			this._transport,
			method,
			opt,
			input,
		);
	}
}