// packages/myso/src/grpc/proto/myso/rpc/v2/move_package_service.client.ts
export interface IMovePackageServiceClient {
	/**
	 * @generated from protobuf rpc: GetPackage(myso.rpc.v2.GetPackageRequest) returns (myso.rpc.v2.GetPackageResponse);
	 */
	getPackage(
		input: GetPackageRequest,
		options?: RpcOptions,
	): UnaryCall<GetPackageRequest, GetPackageResponse>;
	/**
	 * @generated from protobuf rpc: GetDatatype(myso.rpc.v2.GetDatatypeRequest) returns (myso.rpc.v2.GetDatatypeResponse);
	 */
	getDatatype(
		input: GetDatatypeRequest,
		options?: RpcOptions,
	): UnaryCall<GetDatatypeRequest, GetDatatypeResponse>;
	/**
	 * @generated from protobuf rpc: GetFunction(myso.rpc.v2.GetFunctionRequest) returns (myso.rpc.v2.GetFunctionResponse);
	 */
	getFunction(
		input: GetFunctionRequest,
		options?: RpcOptions,
	): UnaryCall<GetFunctionRequest, GetFunctionResponse>;
	/**
	 * @generated from protobuf rpc: ListPackageVersions(myso.rpc.v2.ListPackageVersionsRequest) returns (myso.rpc.v2.ListPackageVersionsResponse);
	 */
	listPackageVersions(
		input: ListPackageVersionsRequest,
		options?: RpcOptions,
	): UnaryCall<ListPackageVersionsRequest, ListPackageVersionsResponse>;
}
/**
 * @generated from protobuf service myso.rpc.v2.MovePackageService
 */
export class MovePackageServiceClient implements IMovePackageServiceClient, ServiceInfo {
	typeName = MovePackageService.typeName;
	methods = MovePackageService.methods;
	options = MovePackageService.options;
	constructor(private readonly _transport: RpcTransport) {}
	/**
	 * @generated from protobuf rpc: GetPackage(myso.rpc.v2.GetPackageRequest) returns (myso.rpc.v2.GetPackageResponse);
	 */
	getPackage(
		input: GetPackageRequest,
		options?: RpcOptions,
	): UnaryCall<GetPackageRequest, GetPackageResponse> {
		const method = this.methods[0],
			opt = this._transport.mergeOptions(options);
		return stackIntercept<GetPackageRequest, GetPackageResponse>(
			'unary',
			this._transport,
			method,
			opt,
			input,
		);
	}
	/**
	 * @generated from protobuf rpc: GetDatatype(myso.rpc.v2.GetDatatypeRequest) returns (myso.rpc.v2.GetDatatypeResponse);
	 */
	getDatatype(
		input: GetDatatypeRequest,
		options?: RpcOptions,
	): UnaryCall<GetDatatypeRequest, GetDatatypeResponse> {
		const method = this.methods[1],
			opt = this._transport.mergeOptions(options);
		return stackIntercept<GetDatatypeRequest, GetDatatypeResponse>(
			'unary',
			this._transport,
			method,
			opt,
			input,
		);
	}
	/**
	 * @generated from protobuf rpc: GetFunction(myso.rpc.v2.GetFunctionRequest) returns (myso.rpc.v2.GetFunctionResponse);
	 */
	getFunction(
		input: GetFunctionRequest,
		options?: RpcOptions,
	): UnaryCall<GetFunctionRequest, GetFunctionResponse> {
		const method = this.methods[2],
			opt = this._transport.mergeOptions(options);
		return stackIntercept<GetFunctionRequest, GetFunctionResponse>(
			'unary',
			this._transport,
			method,
			opt,
			input,
		);
	}
	/**
	 * @generated from protobuf rpc: ListPackageVersions(myso.rpc.v2.ListPackageVersionsRequest) returns (myso.rpc.v2.ListPackageVersionsResponse);
	 */
	listPackageVersions(
		input: ListPackageVersionsRequest,
		options?: RpcOptions,
	): UnaryCall<ListPackageVersionsRequest, ListPackageVersionsResponse> {
		const method = this.methods[3],
			opt = this._transport.mergeOptions(options);
		return stackIntercept<ListPackageVersionsRequest, ListPackageVersionsResponse>(
			'unary',
			this._transport,
			method,
			opt,
			input,
		);
	}
}