// packages/myso/src/grpc/proto/myso/rpc/v2/state_service.client.ts
export interface IStateServiceClient {
	/**
	 * @generated from protobuf rpc: ListDynamicFields(myso.rpc.v2.ListDynamicFieldsRequest) returns (myso.rpc.v2.ListDynamicFieldsResponse);
	 */
	listDynamicFields(
		input: ListDynamicFieldsRequest,
		options?: RpcOptions,
	): UnaryCall<ListDynamicFieldsRequest, ListDynamicFieldsResponse>;
	/**
	 * @generated from protobuf rpc: ListOwnedObjects(myso.rpc.v2.ListOwnedObjectsRequest) returns (myso.rpc.v2.ListOwnedObjectsResponse);
	 */
	listOwnedObjects(
		input: ListOwnedObjectsRequest,
		options?: RpcOptions,
	): UnaryCall<ListOwnedObjectsRequest, ListOwnedObjectsResponse>;
	/**
	 * @generated from protobuf rpc: GetCoinInfo(myso.rpc.v2.GetCoinInfoRequest) returns (myso.rpc.v2.GetCoinInfoResponse);
	 */
	getCoinInfo(
		input: GetCoinInfoRequest,
		options?: RpcOptions,
	): UnaryCall<GetCoinInfoRequest, GetCoinInfoResponse>;
	/**
	 * @generated from protobuf rpc: GetBalance(myso.rpc.v2.GetBalanceRequest) returns (myso.rpc.v2.GetBalanceResponse);
	 */
	getBalance(
		input: GetBalanceRequest,
		options?: RpcOptions,
	): UnaryCall<GetBalanceRequest, GetBalanceResponse>;
	/**
	 * @generated from protobuf rpc: ListBalances(myso.rpc.v2.ListBalancesRequest) returns (myso.rpc.v2.ListBalancesResponse);
	 */
	listBalances(
		input: ListBalancesRequest,
		options?: RpcOptions,
	): UnaryCall<ListBalancesRequest, ListBalancesResponse>;
}
/**
 * @generated from protobuf service myso.rpc.v2.StateService
 */
export class StateServiceClient implements IStateServiceClient, ServiceInfo {
	typeName = StateService.typeName;
	methods = StateService.methods;
	options = StateService.options;
	constructor(private readonly _transport: RpcTransport) {}
	/**
	 * @generated from protobuf rpc: ListDynamicFields(myso.rpc.v2.ListDynamicFieldsRequest) returns (myso.rpc.v2.ListDynamicFieldsResponse);
	 */
	listDynamicFields(
		input: ListDynamicFieldsRequest,
		options?: RpcOptions,
	): UnaryCall<ListDynamicFieldsRequest, ListDynamicFieldsResponse> {
		const method = this.methods[0],
			opt = this._transport.mergeOptions(options);
		return stackIntercept<ListDynamicFieldsRequest, ListDynamicFieldsResponse>(
			'unary',
			this._transport,
			method,
			opt,
			input,
		);
	}
	/**
	 * @generated from protobuf rpc: ListOwnedObjects(myso.rpc.v2.ListOwnedObjectsRequest) returns (myso.rpc.v2.ListOwnedObjectsResponse);
	 */
	listOwnedObjects(
		input: ListOwnedObjectsRequest,
		options?: RpcOptions,
	): UnaryCall<ListOwnedObjectsRequest, ListOwnedObjectsResponse> {
		const method = this.methods[1],
			opt = this._transport.mergeOptions(options);
		return stackIntercept<ListOwnedObjectsRequest, ListOwnedObjectsResponse>(
			'unary',
			this._transport,
			method,
			opt,
			input,
		);
	}
	/**
	 * @generated from protobuf rpc: GetCoinInfo(myso.rpc.v2.GetCoinInfoRequest) returns (myso.rpc.v2.GetCoinInfoResponse);
	 */
	getCoinInfo(
		input: GetCoinInfoRequest,
		options?: RpcOptions,
	): UnaryCall<GetCoinInfoRequest, GetCoinInfoResponse> {
		const method = this.methods[2],
			opt = this._transport.mergeOptions(options);
		return stackIntercept<GetCoinInfoRequest, GetCoinInfoResponse>(
			'unary',
			this._transport,
			method,
			opt,
			input,
		);
	}
	/**
	 * @generated from protobuf rpc: GetBalance(myso.rpc.v2.GetBalanceRequest) returns (myso.rpc.v2.GetBalanceResponse);
	 */
	getBalance(
		input: GetBalanceRequest,
		options?: RpcOptions,
	): UnaryCall<GetBalanceRequest, GetBalanceResponse> {
		const method = this.methods[3],
			opt = this._transport.mergeOptions(options);
		return stackIntercept<GetBalanceRequest, GetBalanceResponse>(
			'unary',
			this._transport,
			method,
			opt,
			input,
		);
	}
	/**
	 * @generated from protobuf rpc: ListBalances(myso.rpc.v2.ListBalancesRequest) returns (myso.rpc.v2.ListBalancesResponse);
	 */
	listBalances(
		input: ListBalancesRequest,
		options?: RpcOptions,
	): UnaryCall<ListBalancesRequest, ListBalancesResponse> {
		const method = this.methods[4],
			opt = this._transport.mergeOptions(options);
		return stackIntercept<ListBalancesRequest, ListBalancesResponse>(
			'unary',
			this._transport,
			method,
			opt,
			input,
		);
	}
}